+++ commodore/compile.py
@@ -1,6 +1,4 @@
 import click
-from kapitan.cached import reset_cache as reset_reclass_cache
-from kapitan.resources import inventory_reclass
 
 from . import git
 from .catalog import fetch_customer_catalog, clean_catalog, update_catalog
@@ -17,13 +15,13 @@ from .dependency_mgmt import (
     fetch_jsonnet_libs,
     fetch_jsonnet_libraries,
     register_components,
-    set_component_overrides,
     jsonnet_dependencies,
 )
 from .helpers import (
     ApiError,
     clean_working_tree,
     kapitan_compile,
+    kapitan_inventory,
 )
 from .postprocess import postprocess_components
 from .refs import update_refs
@@ -116,24 +114,10 @@ def compile(config, cluster_id):
         clean_working_tree(config)
         catalog_repo = _regular_setup(config, cluster_id)
 
-    # Compile kapitan inventory to extract component versions. Component
-    # versions are assumed to be defined in the inventory key
-    # 'parameters.component_versions'
-    reset_reclass_cache()
-    cluster_inventory = inventory_reclass(config.inventory.inventory_dir)["nodes"][
-        config.inventory.bootstrap_target
-    ]
-    versions = cluster_inventory["parameters"].get("component_versions", None)
-    if versions and not config.local:
-        set_component_overrides(config, versions)
-    # Rebuild reclass inventory to use new version of components
-    reset_reclass_cache()
-    kapitan_inventory = inventory_reclass(config.inventory.inventory_dir)["nodes"]
-    cluster_parameters = kapitan_inventory[config.inventory.bootstrap_target][
-        "parameters"
-    ]
+    inventory = kapitan_inventory(config)
+    cluster_parameters = inventory[config.inventory.bootstrap_target]["parameters"]
 
-    # Verify that all aliased components support instantiation after resolving component version overrides.
+    # Verify that all aliased components support instantiation
     config.verify_component_aliases(cluster_parameters)
 
     for component in config.get_components().values():
@@ -155,11 +139,11 @@ def compile(config, cluster_id):
 
     # Generate Kapitan secret references from refs found in inventory
     # parameters
-    update_refs(config, aliases, kapitan_inventory)
+    update_refs(config, aliases, inventory)
 
     kapitan_compile(config, targets, search_paths=[config.vendor_dir])
 
-    postprocess_components(config, kapitan_inventory, components)
+    postprocess_components(config, inventory, components)
 
     update_catalog(config, targets, catalog_repo)